・judgmentを他でも使えるようにフォルダ移動 ・問題表示の際に入力欄が文頭・文末で表示したい場合のロジックを追記
