+++ app/lessons/easy/matcher/test/matcher.ts
@@ -4,7 +4,7 @@ export const matcherTest = () => {
       title: '厳密な等価性',
       question: '「10+20」と「30」が厳密に等しいことをテストしてください。',
       codes: [
-        {code: `test("10+20の結果は30",() => {`},
+        {code: `test('10+20の結果は30',() => {`},
         {code: '  expect(10+20)'},
         {code: 'editor'},
         {code: ';'},
@@ -16,7 +16,7 @@ export const matcherTest = () => {
       title: '配列の等価性のテスト',
       question: '次の配列が等しいことをテストしてください。',
       codes: [
-        {code: `test("2つの配列は等しい", () => {`},
+        {code: `test('2つの配列は等しい', () => {`},
         {code: '  const array1 = [1, 2, 3, 4];'},
         {code: '  const array2 = [1, 2, 3, 4];'},
         {code: '  expect(array1)'},
@@ -30,7 +30,7 @@ export const matcherTest = () => {
       title: '真偽値のテスト',
       question: 'nameがnullではないことをテストしてください。',
       codes: [
-        {code: `test("名前がnullじゃない", () => {`},
+        {code: `test('名前がnullじゃない', () => {`},
         {code: '  const name = "Taro Yamada";'},
         {code: '  expect(name)'},
         {code: 'editor'},
@@ -43,7 +43,7 @@ export const matcherTest = () => {
       title: '数値のテスト',
       question: '年齢が20歳以上であることをテストしてください。',
       codes: [
-        {code: `test("年齢が20歳以上", () => {`},
+        {code: `test('年齢が20歳以上', () => {`},
         {code: '  const age = 25;'},
         {code: '  expect(age)'},
         {code: 'editor'},
@@ -56,7 +56,7 @@ export const matcherTest = () => {
       title: '文字列のテスト',
       question: 'SOZOに「Z」が入っていることをテストしてください。',
       codes: [
-        {code: `test("「Z」が含まれている", () => {`},
+        {code: `test('「Z」が含まれている', () => {`},
         {code: '  expect("SOZO")'},
         {code: 'editor'},
         {code: ';'},
