+++ app/lessons/easy/matcher/test/page.tsx
@@ -3,17 +3,6 @@ import {matcherTest} from './matcher';
 import {useEffect, useState} from 'react';
 import {TestPage} from '@/lib/test-page-format';
 
-type PropsCode = {
-  code: string;
-};
-
-export type PropsItem = {
-  title: string;
-  question: string;
-  codes: PropsCode[];
-  answer: string;
-};
-
 export default function MatcherTestPage() {
   const [testNum, setTestNum] = useState(0);
   const [judge, setJudge] = useState<boolean | null>();
